Add files via upload
In version2, each items will show up according to your points, so that it would`t show up at the begining of the game.

+++ game.h
@@ -71,7 +71,7 @@ public:
     void runMagnet();
     void attract_food();
     void foodMove();
-    static int keeppoints ;
+    static float keeppoints ;
     void revive();
     void help()  const;
     static int shut;
